Windows build fix
svn:sync-xref-src-repo-rev: r83604

+++ src/VBox/Main/include/Performance.h
@@ -1,4 +1,4 @@
-/* $Id: Performance.h 44548 2013-02-05 15:52:49Z aleksey.ilyushin@oracle.com $ */
+/* $Id: Performance.h 44551 2013-02-05 16:21:59Z aleksey.ilyushin@oracle.com $ */
 /** @file
  * VirtualBox Main - Performance Classes declaration.
  */
@@ -536,7 +536,7 @@ namespace pm
         ULONG getScale() { return 1; }
     private:
         com::Utf8Str mShortName;
-        ULONG        mSpeed;
+        uint32_t     mSpeed;
         SubMetric   *mLinkSpeed;
     };
